requirements.txt
attrs==20.3.0
Automat==20.2.0
beautifulsoup4==4.9.3
boto3==1.17.45
botocore==1.20.45
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
filelock==3.0.12
html2text==2020.1.16
hyperlink==21.0.0
idna==2.10
importlib-metadata==3.10.0
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
jsonschema==3.2.0
lxml==4.6.3
parsel==1.6.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pymongo==3.11.3
pyOpenSSL==20.0.1
pyrsistent==0.17.3
python-dateutil==2.8.1
PyYAML==5.4.1
queuelib==1.5.0
requests==2.25.1
requests-file==1.5.1
s3transfer==0.3.6
Scrapy==2.4.1
service-identity==18.1.0
six==1.15.0
soupsieve==2.2.1
tldextract==3.1.0
Twisted==21.2.0
typing-extensions==3.7.4.3
urllib3==1.26.4
w3lib==1.22.0
zipp==3.4.1
zope.interface==5.3.0
